refactor: create hook to centralize minter actions
